Tests/LibWeb/Text/expected/Editing/execCommand-outdent.txt


    <div>foobar</div>
    <ul>
        <li>foo</li>
        <li>bar</li><li></li>
        <li>baz</li>
    </ul>
